DN-11: Split constants.ts into client-safe and server-only files
