+++ .gitignore
@@ -137,3 +137,4 @@ dist
 
 # Finder (MacOS) folder config
 .DS_Store
+.playwright-mcp/
